extract and separate classes for generating marc and writing symphony files
