+++ .rubocop_todo.yml
@@ -1,6 +1,6 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config`
-# on 2022-06-10 21:06:23 UTC using RuboCop version 1.30.1.
+# on 2023-02-17 19:17:57 UTC using RuboCop version 1.45.1.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
@@ -12,8 +12,8 @@ Lint/SuppressedException:
   Exclude:
     - 'lib/tasks/rspec.rake'
 
-# Offense count: 57
-# Configuration parameters: IgnoredMethods, CountRepeatedAttributes.
+# Offense count: 66
+# Configuration parameters: AllowedMethods, AllowedPatterns, CountRepeatedAttributes.
 Metrics/AbcSize:
   Max: 39
 
@@ -22,13 +22,13 @@ Metrics/AbcSize:
 Metrics/ClassLength:
   Max: 114
 
-# Offense count: 11
-# Configuration parameters: IgnoredMethods.
+# Offense count: 12
+# Configuration parameters: AllowedMethods, AllowedPatterns.
 Metrics/CyclomaticComplexity:
   Max: 11
 
-# Offense count: 56
-# Configuration parameters: CountComments, CountAsOne, ExcludedMethods, IgnoredMethods.
+# Offense count: 73
+# Configuration parameters: CountComments, CountAsOne, AllowedMethods, AllowedPatterns.
 Metrics/MethodLength:
   Max: 28
 
@@ -37,15 +37,15 @@ Metrics/MethodLength:
 Metrics/ParameterLists:
   Max: 6
 
-# Offense count: 8
-# Configuration parameters: IgnoredMethods.
+# Offense count: 9
+# Configuration parameters: AllowedMethods, AllowedPatterns.
 Metrics/PerceivedComplexity:
   Max: 12
 
 # Offense count: 11
 Naming/AccessorMethodName:
   Exclude:
-    - 'app/services/update_marc_record_service.rb'
+    - 'app/services/marc_generator.rb'
 
 # Offense count: 9
 # Configuration parameters: EnforcedStyle, AllowedIdentifiers, AllowedPatterns.
@@ -54,7 +54,7 @@ Naming/VariableName:
   Exclude:
     - 'app/services/publish/public_desc_metadata_service.rb'
 
-# Offense count: 13
+# Offense count: 17
 RSpec/AnyInstance:
   Exclude:
     - 'spec/requests/create_collection_spec.rb'
@@ -74,19 +74,18 @@ RSpec/BeforeAfterAll:
   Exclude:
     - 'spec/services/digital_stacks_service_spec.rb'
 
-# Offense count: 8
-# Configuration parameters: Prefixes.
+# Offense count: 7
+# Configuration parameters: Prefixes, AllowedPatterns.
 # Prefixes: when, with, without
 RSpec/ContextWording:
   Exclude:
     - 'spec/services/digital_stacks_service_spec.rb'
-    - 'spec/services/publish/metadata_transfer_service_spec.rb'
     - 'spec/services/publish/public_desc_metadata_service_spec.rb'
     - 'spec/services/publish/public_xml_service_spec.rb'
     - 'spec/services/release_tags/identity_metadata_spec.rb'
     - 'spec/services/thumbnail_service_spec.rb'
 
-# Offense count: 11
+# Offense count: 12
 RSpec/ExpectInHook:
   Exclude:
     - 'spec/services/preservation_ingest_service_spec.rb'
@@ -99,7 +98,7 @@ RSpec/InstanceVariable:
   Exclude:
     - 'spec/services/digital_stacks_service_spec.rb'
 
-# Offense count: 18
+# Offense count: 14
 # Configuration parameters: EnforcedStyle.
 # SupportedStyles: have_received, receive
 RSpec/MessageSpies:
@@ -107,11 +106,11 @@ RSpec/MessageSpies:
     - 'spec/services/digital_stacks_service_spec.rb'
     - 'spec/services/preservation_ingest_service_spec.rb'
     - 'spec/services/preservation_metadata_extractor_spec.rb'
-    - 'spec/services/update_marc_record_service_spec.rb'
     - 'spec/services/version_service_spec.rb'
 
 # Offense count: 6
-# Configuration parameters: IgnoreSharedExamples.
+# Configuration parameters: EnforcedStyle, IgnoreSharedExamples.
+# SupportedStyles: always, named_only
 RSpec/NamedSubject:
   Exclude:
     - 'spec/services/thumbnail_service_spec.rb'
@@ -128,7 +127,7 @@ RSpec/StubbedMock:
     - 'spec/services/preservation_metadata_extractor_spec.rb'
     - 'spec/services/version_service_spec.rb'
 
-# Offense count: 4
+# Offense count: 2
 RSpec/SubjectStub:
   Exclude:
     - 'spec/services/update_marc_record_service_spec.rb'
@@ -170,15 +169,15 @@ Style/Documentation:
     - 'config/application.rb'
     - 'config/initializers/okcomputer.rb'
 
-# Offense count: 11
+# Offense count: 13
 # This cop supports safe autocorrection (--autocorrect).
-# Configuration parameters: Strict, AllowedNumbers.
+# Configuration parameters: Strict, AllowedNumbers, AllowedPatterns.
 Style/NumericLiterals:
   MinDigits: 7
 
 # Offense count: 1
 # This cop supports unsafe autocorrection (--autocorrect-all).
-# Configuration parameters: EnforcedStyle, IgnoredMethods.
+# Configuration parameters: EnforcedStyle, AllowedMethods, AllowedPatterns.
 # SupportedStyles: predicate, comparison
 Style/NumericPredicate:
   Exclude:
@@ -221,9 +220,9 @@ Style/SymbolArray:
     - 'config/routes.rb'
     - 'lib/tasks/rspec.rake'
 
-# Offense count: 403
+# Offense count: 428
 # This cop supports safe autocorrection (--autocorrect).
-# Configuration parameters: AllowHeredoc, AllowURI, URISchemes, IgnoreCopDirectives, AllowedPatterns, IgnoredPatterns.
+# Configuration parameters: AllowHeredoc, AllowURI, URISchemes, IgnoreCopDirectives, AllowedPatterns.
 # URISchemes: http, https
 Layout/LineLength:
-  Max: 191
+  Max: 300
