+++ app/controllers/objects_controller.rb
@@ -129,7 +129,7 @@ class ObjectsController < ApplicationController
   end
 
   def update_marc_record
-    UpdateMarcRecordService.new(@cocina_object, thumbnail_service: ThumbnailService.new(@cocina_object)).update
+    UpdateMarcRecordService.update(@cocina_object, thumbnail_service: ThumbnailService.new(@cocina_object))
     head :created
   end
